+++ bbclib/libs/bbclib_relation.py
@@ -23,13 +23,16 @@ sys.path.append(os.path.join(current_dir, "../.."))
 from bbclib.libs import bbclib_utils
 from bbclib.libs.bbclib_pointer import BBcPointer
 from bbclib.libs.bbclib_asset import BBcAsset
+from bbclib.libs.bbclib_asset_raw import BBcAssetRaw
+from bbclib.libs.bbclib_asset_hash import BBcAssetHash
 import bbclib
 from bbclib import id_length_conf
 
 
 class BBcRelation:
     """Relation part in a transaction"""
-    def __init__(self, asset_group_id=None, id_length=None):
+    def __init__(self, asset_group_id=None, id_length=None, version=2):
+        self.version = version
         if id_length is not None:
             bbclib.configure_id_length_all(id_length)
         if asset_group_id is not None:
@@ -38,6 +41,8 @@ class BBcRelation:
             self.asset_group_id = None
         self.pointers = list()
         self.asset = None
+        self.asset_raw = None
+        self.asset_hash = None
 
     def __str__(self):
         ret =  "  asset_group_id: %s\n" % bbclib_utils.str_binary(self.asset_group_id)
@@ -46,10 +51,15 @@ class BBcRelation:
             for i, pt in enumerate(self.pointers):
                 ret += "   [%d]\n" % i
                 ret += str(pt)
-        ret += str(self.asset)
+        if self.asset is not None:
+            ret += str(self.asset)
+        if self.asset_raw is not None:
+            ret += str(self.asset_raw)
+        if self.asset_hash is not None:
+            ret += str(self.asset_hash)
         return ret
 
-    def add(self, asset_group_id=None, asset=None, pointer=None):
+    def add(self, asset_group_id=None, asset=None, asset_raw=None, asset_hash=None, pointer=None):
         """Add parts"""
         if asset_group_id is not None:
             self.asset_group_id = asset_group_id[:id_length_conf["asset_group_id"]]
@@ -60,6 +70,10 @@ class BBcRelation:
                 self.pointers.append(pointer)
         if asset is not None:
             self.asset = asset
+        if asset_raw is not None:
+            self.asset_raw = asset_raw
+        if asset_hash is not None:
+            self.asset_hash = asset_hash
         return True
 
     def pack(self):
@@ -80,17 +94,32 @@ class BBcRelation:
             dat.extend(ast)
         else:
             dat.extend(bbclib_utils.to_4byte(0))
+        if self.version >= 3:
+            if self.asset_raw is not None:
+                ast = self.asset_raw.pack()
+                dat.extend(bbclib_utils.to_4byte(len(ast)))
+                dat.extend(ast)
+            else:
+                dat.extend(bbclib_utils.to_4byte(0))
+            if self.asset_hash is not None:
+                ast = self.asset_hash.pack()
+                dat.extend(bbclib_utils.to_4byte(len(ast)))
+                dat.extend(ast)
+            else:
+                dat.extend(bbclib_utils.to_4byte(0))
         return bytes(dat)
 
-    def unpack(self, data):
+    def unpack(self, data, version=2):
         """Unpack data into transaction object
 
         Args:
             data (bytes): packed binary data
+            version (int): version of the data format
         Returns:
             bool: True if successful
         """
         ptr = 0
+        self.version = version
         data_size = len(data)
         try:
             ptr, self.asset_group_id = bbclib_utils.get_bigint(ptr, data)
@@ -113,6 +142,23 @@ class BBcRelation:
                 ptr, astdata = bbclib_utils.get_n_bytes(ptr, astsize, data)
                 if not self.asset.unpack(astdata):
                     return False
+
+            if version >= 3:
+                self.asset_raw = None
+                ptr, astsize = bbclib_utils.get_n_byte_int(ptr, 4, data)
+                if astsize > 0:
+                    self.asset_raw = BBcAssetRaw()
+                    ptr, astdata = bbclib_utils.get_n_bytes(ptr, astsize, data)
+                    if not self.asset_raw.unpack(astdata):
+                        return False
+                self.asset_hash = None
+                ptr, astsize = bbclib_utils.get_n_byte_int(ptr, 4, data)
+                if astsize > 0:
+                    self.asset_hash = BBcAssetHash()
+                    ptr, astdata = bbclib_utils.get_n_bytes(ptr, astsize, data)
+                    if not self.asset_hash.unpack(astdata):
+                        return False
+
         except:
             return False
         return True